split WasmModule to 3 classes
